Agent_d1/requirements.txt
openai>=1.0.0
numpy
python-dotenv
PyPDF2
